Support for multiple Eqlinks per tile and activity
