+++ src/Core/main.cpp
@@ -28,6 +28,7 @@
 #include "PowerProfile.h"
 #include "GcCrashDialog.h" // for versionHTML
 #include "OverviewItems.h"
+#include "OverviewEquipmentItems.h"
 
 #include <QApplication>
 #include <QtGui>
@@ -608,6 +609,7 @@ main(int argc, char *argv[])
 
         // initialize Overview Items once the translator is installed
         OverviewItemConfig::registerItems();
+        OverviewEquipmentItemConfig::registerItems();
 
         // initialise the trainDB
         trainDB = new TrainDB(home);
