+++ src/Gui/AthleteTab.cpp
@@ -123,10 +123,7 @@ AthleteTab::~AthleteTab()
 void
 AthleteTab::close()
 {
-    analysisView->saveState();
-    homeView->saveState();
-    trainView->saveState();
-    diaryView->saveState();
+    // when the views are deleted they save their state
 
     analysisView->close();
     homeView->close();
