+++ src/Gui/Views.cpp
@@ -63,7 +63,6 @@ AnalysisView::~AnalysisView()
 {
     appsettings->setValue(GC_SETTINGS_MAIN_SIDEBAR "analysis", _sidebar);
     delete analSidebar;
-    //delete hw; tabview deletes after save state
 }
 
 void
@@ -152,7 +151,6 @@ DiaryView::~DiaryView()
 {
     appsettings->setValue(GC_SETTINGS_MAIN_SIDEBAR "diary", _sidebar);
     delete diarySidebar;
-    //delete hw; tabview deletes after save state
 }
 
 void
@@ -207,7 +205,6 @@ TrendsView::~TrendsView()
 {
     appsettings->setValue(GC_SETTINGS_MAIN_SIDEBAR "trend", _sidebar);
     delete sidebar;
-    //delete hw; tabview deletes after save state
 }
 
 void
@@ -322,7 +319,6 @@ TrainView::~TrainView()
 {
     appsettings->setValue(GC_SETTINGS_MAIN_SIDEBAR "train", _sidebar);
     delete trainTool;
-    //delete hw; tabview deletes after save state
 }
 
 void
@@ -369,7 +365,7 @@ EquipView::EquipView(Context* context, QStackedWidget* controls) :
 
 EquipView::~EquipView()
 {
-    saveState(); // writes equipment-perspectives.xml
+    // No sidebar to delete
 }
 
 bool
