+++ src/Resources/charts/overview-equipment.gchart
@@ -4,13 +4,14 @@
     "VIEW": "equipment",
     "TYPE": "51",
     "PROPERTIES": {
-      "title": "EquipmentLink ",
+      "title": "Equipment Overview ",
       "subtitle": " ",
       "widthFactor": "2",
       "heightFactor": "2",
       "style": "0",
       "resizable": "0",
-      "config": "{\n  \"version\":\"2.0\",\n  \"CHARTS\":[\n    { \"type\":115,\"span\":1,\"deep\":14,\"column\":0,\"order\":0,\"color\":\"#010161\",\"showAthleteActivities\":\"1\",\"datafilter\":\" \",\"name\":\"Bike1 Summary\" },\n    { \"type\":114,\"span\":1,\"deep\":13,\"column\":1,\"order\":7,\"color\":\"#010161\",\"nonGCDistanceScaled\":\"0\",\"nonGCElevationScaled\":\"0\",\"repDistance\":\"0\",\"repElevation\":\"0\",\"startSet\":\"0\",\"endSet\":\"0\",\"notes\":\"Frame notes here\",\"datafilter\":\" \",\"name\":\"Frame\" },\n    { \"type\":116,\"span\":1,\"deep\":7,\"column\":0,\"order\":1,\"color\":\"#010161\",\"notes\":\"Equipment notes here\",\"datafilter\":\" \",\"name\":\"Notes\" },\n    { \"type\":114,\"span\":1,\"deep\":13,\"column\":2,\"order\":0,\"color\":\"#010161\",\"nonGCDistanceScaled\":\"0\",\"nonGCElevationScaled\":\"0\",\"repDistance\":\"0\",\"repElevation\":\"0\",\"startSet\":\"0\",\"endSet\":\"0\",\"notes\":\"Wheelset notes here\",\"datafilter\":\" \",\"name\":\"Wheelset\" }\n  ]\n}\n ",
+      "config": "{\n  \"version\":\"2.0\",\n  \"widths\":[ 1000, 1100, 1150, 1150, 1200, 1200, 1200, 1200, 1200, 1200 ],\n  \"CHARTS\":[\n    { \"type\":114,\"span\":1,\"deep\":12,\"column\":1,\"order\":24,\"color\":\"#010161\",\"metric\":\"0\",\"nonGCDistanceScaled\":\"0\",\"nonGCElevationScaled\":\"0\",\"repDistance\":\"0\",\"repElevation\":\"0\",\"eqUseList\":[{\"endDate\":\"\",\"endSet\":\"0\",\"eqLink\":\"Bike1\",\"startDate\":\"Mon Sep 1 2014\",\"startSet\":\"1\"}],\"notes\":\"\",\"datafilter\":\" \",\"name\":\"Frame and Fork\" },\n    { \"type\":115,\"span\":1,\"deep\":14,\"column\":0,\"order\":16,\"color\":\"#010161\",\"eqLink\":\"Bike 1\",\"showAthleteActivities\":\"1\",\"datafilter\":\" \",\"name\":\"Bike1 Summary\" },\n    { \"type\":116,\"span\":1,\"deep\":6,\"column\":0,\"order\":24,\"color\":\"#010161\",\"notes\":\"Notes here\",\"datafilter\":\" \",\"name\":\"Notes\" },\n    { \"type\":114,\"span\":1,\"deep\":13,\"column\":1,\"order\":25,\"color\":\"#010161\",\"metric\":\"0\",\"nonGCDistanceScaled\":\"0\",\"nonGCElevationScaled\":\"0\",\"repDistance\":\"0\",\"repElevation\":\"0\",\"eqUseList\":[{\"endDate\":\"\",\"endSet\":\"0\",\"eqLink\":\"Bike1\",\"startDate\":\"Fri May 3 2024\",\"startSet\":\"1\"}],\"notes\":\"\",\"datafilter\":\" \",\"name\":\"Brake blocks\/Pads\" },\n    { \"type\":114,\"span\":1,\"deep\":13,\"column\":2,\"order\":81,\"color\":\"#010161\",\"metric\":\"0\",\"nonGCDistanceScaled\":\"0\",\"nonGCElevationScaled\":\"0\",\"repDistance\":\"0\",\"repElevation\":\"0\",\"eqUseList\":[{\"endDate\":\"\",\"endSet\":\"0\",\"eqLink\":\"Bike1\",\"startDate\":\"Fri Oct 20 2023\",\"startSet\":\"1\"}],\"notes\":\"\",\"datafilter\":\" \",\"name\":\"Front Chainrings\" },\n    { \"type\":114,\"span\":1,\"deep\":12,\"column\":2,\"order\":82,\"color\":\"#010161\",\"metric\":\"0\",\"nonGCDistanceScaled\":\"0\",\"nonGCElevationScaled\":\"0\",\"repDistance\":\"0\",\"repElevation\":\"0\",\"eqUseList\":[{\"endDate\":\"\",\"endSet\":\"0\",\"eqLink\":\"Bike1\",\"startDate\":\"Sat Mar 11 2023\",\"startSet\":\"1\"}],\"notes\":\"\",\"datafilter\":\" \",\"name\":\"Cassette\" },\n    { \"type\":114,\"span\":1,\"deep\":12,\"column\":2,\"order\":83,\"color\":\"#010161\",\"metric\":\"0\",\"nonGCDistanceScaled\":\"0\",\"nonGCElevationScaled\":\"0\",\"repDistance\":\"0\",\"repElevation\":\"0\",\"eqUseList\":[{\"endDate\":\"\",\"endSet\":\"0\",\"eqLink\":\"Bike1\",\"startDate\":\"Fri Oct 20 2023\",\"startSet\":\"1\"}],\"notes\":\"\",\"datafilter\":\" \",\"name\":\"Chain\" },\n    { \"type\":114,\"span\":1,\"deep\":15,\"column\":3,\"order\":0,\"color\":\"#010161\",\"metric\":\"0\",\"nonGCDistanceScaled\":\"0\",\"nonGCElevationScaled\":\"0\",\"repDistance\":\"0\",\"repElevation\":\"0\",\"eqUseList\":[{\"endDate\":\"\",\"endSet\":\"0\",\"eqLink\":\"Bike1\",\"startDate\":\"Wed May 3 2023\",\"startSet\":\"1\"},{\"endDate\":\"Tue May 2 2023\",\"endSet\":\"1\",\"eqLink\":\"Bike2\",\"startDate\":\"Fri Jan 1 2010\",\"startSet\":\"1\"}],\"notes\":\"\",\"datafilter\":\" \",\"name\":\"Wheelset\" },\n    { \"type\":114,\"span\":1,\"deep\":12,\"column\":3,\"order\":0,\"color\":\"#010161\",\"metric\":\"0\",\"nonGCDistanceScaled\":\"0\",\"nonGCElevationScaled\":\"0\",\"repDistance\":\"0\",\"repElevation\":\"0\",\"eqUseList\":[{\"endDate\":\"\",\"endSet\":\"0\",\"eqLink\":\"Bike1\",\"startDate\":\"Fri Aug 23 2024\",\"startSet\":\"1\"}],\"notes\":\"\",\"datafilter\":\" \",\"name\":\"Front Tyre\" },\n    { \"type\":114,\"span\":1,\"deep\":13,\"column\":3,\"order\":0,\"color\":\"#010161\",\"metric\":\"0\",\"nonGCDistanceScaled\":\"0\",\"nonGCElevationScaled\":\"0\",\"repDistance\":\"0\",\"repElevation\":\"0\",\"eqUseList\":[{\"endDate\":\"\",\"endSet\":\"0\",\"eqLink\":\"Bike1\",\"startDate\":\"Sat Dec 2 2023\",\"startSet\":\"1\"}],\"notes\":\"\",\"datafilter\":\" \",\"name\":\"Back Tyre\" }\n  ]\n}\n ",
+      "minimumColumns": "5",
       "__LAST__": "1"
     }
   }
