+++ src/src.pro
@@ -582,8 +582,8 @@ QT += charts opengl
 
 # Dashboard uses qt charts, so needs at least Qt 5.7
 DEFINES += GC_HAVE_OVERVIEW
-HEADERS += Gui/ChartSpace.h Charts/OverviewItems.h Charts/Overview.h Charts/OverviewEquipment.h
-SOURCES += Gui/ChartSpace.cpp Charts/OverviewItems.cpp Charts/Overview.cpp Charts/OverviewEquipment.cpp
+HEADERS += Gui/ChartSpace.h Charts/OverviewItems.h Charts/Overview.h Charts/OverviewEquipment.h Charts/OverviewEquipmentItems.h Charts/EquipmentCalculator.h
+SOURCES += Gui/ChartSpace.cpp Charts/OverviewItems.cpp Charts/Overview.cpp Charts/OverviewEquipment.cpp Charts/OverviewEquipmentItems.cpp Charts/EquipmentCalculator.cpp
 
 # generic chart
 DEFINES += GC_HAVE_GENERIC
